+++ src/index.ts
@@ -4,4 +4,7 @@ export * from './agent.schema.js';
 export * from './api.schema.js';
 export * from './cool.schema.js';
 export * from './device.schema.js';
+export * from './patch-operation.schema.js';
 export * from './screen-details.schema.js';
+export * from './sqlite-date.schema.js';
+export * from './uri.schema.js';